code improvement duplications Signed-off-by: dbulahov <bulahovdenis@gmail.com>
