+++ xmla/ws/jakarta.model.xmla/src/main/java/org/eclipse/daanse/xmla/ws/jakarta/model/xmla/engine300/CalculationPropertiesVisualizationProperties.java
@@ -14,7 +14,6 @@
 package org.eclipse.daanse.xmla.ws.jakarta.model.xmla.engine300;
 
 import java.io.Serializable;
-import java.math.BigInteger;
 
 import jakarta.xml.bind.annotation.XmlAccessType;
 import jakarta.xml.bind.annotation.XmlAccessorType;
@@ -22,100 +21,15 @@ import jakarta.xml.bind.annotation.XmlElement;
 import jakarta.xml.bind.annotation.XmlType;
 
 @XmlAccessorType(XmlAccessType.FIELD)
-@XmlType(name = "CalculationPropertiesVisualizationProperties", propOrder = {"folderPosition", "contextualNameRule",
-    "alignment", "isFolderDefault", "isRightToLeft", "sortDirection", "units", "width", "isDefaultMeasure",
-    "defaultDetailsPosition", "sortPropertiesPosition", "isSimpleMeasure"})
-public class CalculationPropertiesVisualizationProperties implements Serializable {
+@XmlType(name = "CalculationPropertiesVisualizationProperties", propOrder = {"isDefaultMeasure",
+    "isSimpleMeasure"})
+public class CalculationPropertiesVisualizationProperties extends AbstractProperties implements Serializable {
 
     private static final long serialVersionUID = 1L;
-    @XmlElement(name = "FolderPosition", defaultValue = "-1")
-    protected BigInteger folderPosition;
-    @XmlElement(name = "ContextualNameRule", defaultValue = "None")
-    protected String contextualNameRule;
-    @XmlElement(name = "Alignment", defaultValue = "Default")
-    protected String alignment;
-    @XmlElement(name = "IsFolderDefault", defaultValue = "false")
-    protected Boolean isFolderDefault;
-    @XmlElement(name = "IsRightToLeft", defaultValue = "false")
-    protected Boolean isRightToLeft;
-    @XmlElement(name = "SortDirection", defaultValue = "Default")
-    protected String sortDirection;
-    @XmlElement(name = "Units", defaultValue = "")
-    protected String units;
-    @XmlElement(name = "Width", defaultValue = "-1")
-    protected BigInteger width;
     @XmlElement(name = "IsDefaultMeasure", defaultValue = "false")
     protected Boolean isDefaultMeasure;
-    @XmlElement(name = "DefaultDetailsPosition", defaultValue = "-1")
-    protected BigInteger defaultDetailsPosition;
-    @XmlElement(name = "SortPropertiesPosition", defaultValue = "-1")
-    protected BigInteger sortPropertiesPosition;
     @XmlElement(name = "IsSimpleMeasure", defaultValue = "false")
     protected Boolean isSimpleMeasure;
-
-    public BigInteger getFolderPosition() {
-        return folderPosition;
-    }
-
-    public void setFolderPosition(BigInteger folderPosition) {
-        this.folderPosition = folderPosition;
-    }
-
-    public String getContextualNameRule() {
-        return contextualNameRule;
-    }
-
-    public void setContextualNameRule(String contextualNameRule) {
-        this.contextualNameRule = contextualNameRule;
-    }
-
-    public String getAlignment() {
-        return alignment;
-    }
-
-    public void setAlignment(String alignment) {
-        this.alignment = alignment;
-    }
-
-    public Boolean getFolderDefault() {
-        return isFolderDefault;
-    }
-
-    public void setFolderDefault(Boolean folderDefault) {
-        isFolderDefault = folderDefault;
-    }
-
-    public Boolean getRightToLeft() {
-        return isRightToLeft;
-    }
-
-    public void setRightToLeft(Boolean rightToLeft) {
-        isRightToLeft = rightToLeft;
-    }
-
-    public String getSortDirection() {
-        return sortDirection;
-    }
-
-    public void setSortDirection(String sortDirection) {
-        this.sortDirection = sortDirection;
-    }
-
-    public String getUnits() {
-        return units;
-    }
-
-    public void setUnits(String units) {
-        this.units = units;
-    }
-
-    public BigInteger getWidth() {
-        return width;
-    }
-
-    public void setWidth(BigInteger width) {
-        this.width = width;
-    }
 
     public Boolean getDefaultMeasure() {
         return isDefaultMeasure;
@@ -123,22 +37,6 @@ public class CalculationPropertiesVisualizationProperties implements Serializabl
 
     public void setDefaultMeasure(Boolean defaultMeasure) {
         isDefaultMeasure = defaultMeasure;
-    }
-
-    public BigInteger getDefaultDetailsPosition() {
-        return defaultDetailsPosition;
-    }
-
-    public void setDefaultDetailsPosition(BigInteger defaultDetailsPosition) {
-        this.defaultDetailsPosition = defaultDetailsPosition;
-    }
-
-    public BigInteger getSortPropertiesPosition() {
-        return sortPropertiesPosition;
-    }
-
-    public void setSortPropertiesPosition(BigInteger sortPropertiesPosition) {
-        this.sortPropertiesPosition = sortPropertiesPosition;
     }
 
     public Boolean getSimpleMeasure() {
