+++ xmla/ws/jakarta.model.xmla/src/main/java/org/eclipse/daanse/xmla/ws/jakarta/model/xmla/engine300/DimensionAttributeVisualizationProperties.java
@@ -13,7 +13,6 @@
  */
 package org.eclipse.daanse.xmla.ws.jakarta.model.xmla.engine300;
 
-import java.io.Serializable;
 import java.math.BigInteger;
 
 import jakarta.xml.bind.annotation.XmlAccessType;
@@ -22,35 +21,13 @@ import jakarta.xml.bind.annotation.XmlElement;
 import jakarta.xml.bind.annotation.XmlType;
 
 @XmlAccessorType(XmlAccessType.FIELD)
-@XmlType(name = "DimensionAttributeVisualizationProperties", propOrder = {"folderPosition", "contextualNameRule",
-    "alignment", "isFolderDefault", "isRightToLeft", "sortDirection", "units", "width", "defaultDetailsPosition",
-    "commonIdentifierPosition", "sortPropertiesPosition", "displayKeyPosition", "isDefaultImage",
-    "defaultAggregateFunction"})
-public class DimensionAttributeVisualizationProperties implements Serializable {
+@XmlType(name = "DimensionAttributeVisualizationProperties", propOrder = {"commonIdentifierPosition",
+    "displayKeyPosition", "isDefaultImage", "defaultAggregateFunction"})
+public class DimensionAttributeVisualizationProperties extends AbstractProperties {
 
     private static final long serialVersionUID = 1L;
-    @XmlElement(name = "FolderPosition", defaultValue = "-1")
-    protected BigInteger folderPosition;
-    @XmlElement(name = "ContextualNameRule", defaultValue = "None")
-    protected String contextualNameRule;
-    @XmlElement(name = "Alignment", defaultValue = "Default")
-    protected String alignment;
-    @XmlElement(name = "IsFolderDefault", defaultValue = "false")
-    protected Boolean isFolderDefault;
-    @XmlElement(name = "IsRightToLeft", defaultValue = "false")
-    protected Boolean isRightToLeft;
-    @XmlElement(name = "SortDirection", defaultValue = "Default")
-    protected String sortDirection;
-    @XmlElement(name = "Units", defaultValue = "")
-    protected String units;
-    @XmlElement(name = "Width", defaultValue = "-1")
-    protected BigInteger width;
-    @XmlElement(name = "DefaultDetailsPosition", defaultValue = "-1")
-    protected BigInteger defaultDetailsPosition;
     @XmlElement(name = "CommonIdentifierPosition", defaultValue = "-1")
     protected BigInteger commonIdentifierPosition;
-    @XmlElement(name = "SortPropertiesPosition", defaultValue = "-1")
-    protected BigInteger sortPropertiesPosition;
     @XmlElement(name = "DisplayKeyPosition", defaultValue = "-1")
     protected BigInteger displayKeyPosition;
     @XmlElement(name = "IsDefaultImage", defaultValue = "false")
@@ -58,92 +35,12 @@ public class DimensionAttributeVisualizationProperties implements Serializable {
     @XmlElement(name = "DefaultAggregateFunction", defaultValue = "Default")
     protected String defaultAggregateFunction;
 
-    public BigInteger getFolderPosition() {
-        return folderPosition;
-    }
-
-    public void setFolderPosition(BigInteger folderPosition) {
-        this.folderPosition = folderPosition;
-    }
-
-    public String getContextualNameRule() {
-        return contextualNameRule;
-    }
-
-    public void setContextualNameRule(String contextualNameRule) {
-        this.contextualNameRule = contextualNameRule;
-    }
-
-    public String getAlignment() {
-        return alignment;
-    }
-
-    public void setAlignment(String alignment) {
-        this.alignment = alignment;
-    }
-
-    public Boolean getFolderDefault() {
-        return isFolderDefault;
-    }
-
-    public void setFolderDefault(Boolean folderDefault) {
-        isFolderDefault = folderDefault;
-    }
-
-    public Boolean getRightToLeft() {
-        return isRightToLeft;
-    }
-
-    public void setRightToLeft(Boolean rightToLeft) {
-        isRightToLeft = rightToLeft;
-    }
-
-    public String getSortDirection() {
-        return sortDirection;
-    }
-
-    public void setSortDirection(String sortDirection) {
-        this.sortDirection = sortDirection;
-    }
-
-    public String getUnits() {
-        return units;
-    }
-
-    public void setUnits(String units) {
-        this.units = units;
-    }
-
-    public BigInteger getWidth() {
-        return width;
-    }
-
-    public void setWidth(BigInteger width) {
-        this.width = width;
-    }
-
-    public BigInteger getDefaultDetailsPosition() {
-        return defaultDetailsPosition;
-    }
-
-    public void setDefaultDetailsPosition(BigInteger defaultDetailsPosition) {
-        this.defaultDetailsPosition = defaultDetailsPosition;
-    }
-
     public BigInteger getCommonIdentifierPosition() {
         return commonIdentifierPosition;
     }
 
     public void setCommonIdentifierPosition(BigInteger commonIdentifierPosition) {
         this.commonIdentifierPosition = commonIdentifierPosition;
-    }
-
-    public BigInteger getSortPropertiesPosition() {
-        return sortPropertiesPosition;
-    }
-
-    public void setSortPropertiesPosition(BigInteger sortPropertiesPosition) {
-        this.sortPropertiesPosition = sortPropertiesPosition;
     }
 
     public BigInteger getDisplayKeyPosition() {
