+++ xmla/ws/jakarta.model.xmla/src/main/java/org/eclipse/daanse/xmla/ws/jakarta/model/xmla/xmla/AbstractAggregation.java
@@ -1,0 +1,71 @@
+/*
+ * Copyright (c) 2023 Contributors to the Eclipse Foundation.
+ *
+ * This program and the accompanying materials are made
+ * available under the terms of the Eclipse Public License 2.0
+ * which is available at https://www.eclipse.org/legal/epl-2.0/
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Contributors:
+ *   SmartCity Jena - initial
+ *   Stefan Bischof (bipolis.org) - initial
+ */
+package org.eclipse.daanse.xmla.ws.jakarta.model.xmla.xmla;
+
+import jakarta.xml.bind.annotation.XmlAccessType;
+import jakarta.xml.bind.annotation.XmlAccessorType;
+import jakarta.xml.bind.annotation.XmlElement;
+import jakarta.xml.bind.annotation.XmlElementWrapper;
+import jakarta.xml.bind.annotation.XmlSeeAlso;
+import jakarta.xml.bind.annotation.XmlType;
+
+import java.util.List;
+
+@XmlAccessorType(XmlAccessType.FIELD)
+@XmlType(name = "", propOrder = { "annotations" })
+@XmlSeeAlso({Aggregation.class, AggregationInstance.class})
+public abstract class AbstractAggregation {
+
+    @XmlElement(name = "ID")
+    protected String id;
+    @XmlElement(name = "Name", required = true)
+    protected String name;
+    @XmlElement(name = "Annotation", type = Annotation.class)
+    @XmlElementWrapper(name = "Annotations")
+    protected List<Annotation> annotations;
+    @XmlElement(name = "Description")
+    protected String description;
+
+    public String getID() {
+        return id;
+    }
+
+    public void setID(String id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public List<Annotation> getAnnotations() {
+        return annotations;
+    }
+
+    public void setAnnotations(List<Annotation> value) {
+        this.annotations = value;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+}
